feat: refactor Navbar component for responsive design and update import paths
