+++ Game-Site/src/Components/Navbar.jsx
@@ -1,22 +1,31 @@
+import { useRef } from "react";
 import { FaBars, FaTimes } from "react-icons/fa";
 
 
 function Navbar(){
-    retrun(
+    const navRef = useRef();
+
+    const showNavbar = () => {
+        navRef.current.classList.toggle("reponsive_nav");
+    }
+
+    return(
         <header>
-            <nav>
+            <nav ref={navRef}>
                 <h1>Logo</h1>
                 <a href="#">Home</a>
                 <a href="#">About Us</a>
                 <a href="#">Blog</a>
                 <a href="#">About Us</a>
-                <button>
-                    <FaBars />
+                <button className={"nav-btn nav-close-btn"} onClick={showNavbar}>
+                    <FaTimes />
                 </button>
             </nav>
-            <button>
-                <FaTimes />
+            <button className={"nav-btn"} onClick={showNavbar}>
+                <FaBars />
             </button>
         </header>
     );
 }
+
+export default Navbar;
